* Setting Folders Namepaces. * Moved AutoMapper (Book_11)
