+++ BookApplication.Services/BookApplication.Services.csproj
@@ -7,10 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="13.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\BookApplication.Data\BookApplication.Data.csproj" />
   </ItemGroup>
